Add solution for number of steps to reduce a number to zero problem
Added a Python solution for calculating the number of steps required to reduce a given number to zero.

Method used: Repeatedly divide the number by 2 if it is even, or subtract 1 if it is odd, until the number becomes zero. Count the number of steps taken.

Time complexity is O(log n) and space complexity is O(1).