Add solution for check if a string is an Acronym of Words problem
Implemented a Python solution to determine if a given string is an acronym of an array of words. 

Method used: Concatenate the first character of each word in the array and compare the resulting string with the given string.

Time complexity is O(n) and space complexity is O(n).