Add solution for find the integer added to array I problem
Implemented a Python solution to find the integer 'x' that, when added to each element of 'nums1', makes it equal to 'nums2'. 

Method used: Sort both arrays and find the difference between the first elements of the sorted arrays. 

Time complexity is O(n log n) and space complexity is O(1).