Add solution for determine color of a chessboard square problem
Added a Python solution to determine the color of a chessboard square based on its coordinates. 

Method used: Convert the letter to a number and check if the sum of the letter’s number and the digit is even or odd. 

Time complexity is O(1) and space complexity is O(1).