Add solution for unique number of occurrences problem
Added a Python solution to check if the number of occurrences of each value in an array is unique. 

Method used: Use a Counter to count occurrences and compare the length of the list of counts to the length of the set of counts.

Time complexity is O(n) and space complexity is O(n).