Add solution for ransom note problem
Added a Python solution to determine if a ransom note can be constructed from a given magazine.

Method used: Use the Counter class from the collections module to count the frequency of each letter in both 'ransomNote' and magazine. Compare the counts to determine if 'ransomNote' can be constructed from magazine.

Time complexity is O(n + m) and space complexity is O(1).