Add solution for greatest common divisor of strings problem
Added a Python solution to find the greatest common divisor of two given strings. 

Method used: Check if concatenations of 'str1' and 'str2' in both orders are equal. Return the substring of 'str1' up to the greatest common divisor of their lengths. 

Time complexity is O(n + m) and space complexity is O(1).