Add solution for isomorphic strings problem
Added a Python solution to determine if two given strings are isomorphic. 

Method used: Use sets to check if the zipped pairs of characters in 's' and 't' have the same length as the sets of unique characters in 's' and 't'. 

Time complexity is O(n) and space complexity is O(n).