Add solution for find occurrences of an element in an array problem
Added a Python solution to find the index of the queries[i]th occurrence of an integer 'x' in an array 'nums'. 

Method used: Gather indices of occurrences of 'x' in 'nums' and return the appropriate index for each query or -1 if there are fewer occurrences than requested.

Time complexity is O(n + q) and space complexity is O(n).