Add solution for search in rotated sorted array problem
Added a Python solution to find the index of a given target value in a possibly rotated sorted integer array 'nums'.

Method used: Check if the target is present in the array using 'in' and return its index with 'index' method. If the target is not found, return -1.

Time complexity is O(n) and space complexity is O(1).