Add solution for get the size of a dataframe problem
Added a Python solution to calculate and display the number of rows and columns of the players DataFrame.

Method used: Use the 'shape' attribute in pandas to get the dimensions of the DataFrame. Return the dimensions as an array [number of rows, number of columns].

Time complexity is O(1) and space complexity is O(1).