Add solution for replace employee id with the unique identifier problem
Added a Python solution to show the unique ID of each user. If a user does not have a unique ID, display null.

Method used: Merge the 'employees' and 'employee_uni' tables on the 'id' column using a left join and select the columns 'unique_id' and 'name' from the merged DataFrame.

Time complexity is O(n) and space complexity is O(n).