Add solution for bank account summary II problem
Added a Python solution to report the name and balance of users with a balance higher than 10000. The balance of an account is equal to the sum of the amounts of all transactions involving that account.

Method used: Group the transactions by account and sum the amounts to get the account balance. Filter the accounts with a balance higher than 10000. Merge this result with the users table to get the names and balances.

Time complexity is O(n) and space complexity is O(n).