Add solution for combine two tables problem
Added a Python solution to report the first name, last name, city, and state of each person in the 'Person' table. If the address of a 'personId' is not present in the 'Address' table, report null instead.

Method used: Merge the 'Person' and 'Address' tables on 'personId' using a left join and select the columns 'firstName', 'lastName', 'city', and 'state' from the merged DataFrame.

Time complexity is O(n) and space complexity is O(n).