Add solution for find customer referee problem
Added a Python solution to find the names of customers that are not referred by the customer with 'id' = 2. Return the result table in any order.

Method used: Filter the DataFrame to include only the rows where 'referee_id' is not equal to 2 or is NULL. Select the 'name' column from the filtered DataFrame.

Time complexity is O(n) and space complexity is O(n).