Add solution for percentage of users attended a contest problem
Added a Python solution to find the percentage of users registered in each contest, rounded to two decimals. Return the result table ordered by 'percentage' in descending order, with 'contest_id' used as a tiebreaker in ascending order.

Method used: Merge the 'users' DataFrame with the 'register' DataFrame on 'user_id'. Group by 'contest_id' and count the number of users. Calculate the percentage of users and sort the results.

Time complexity is O(n) and space complexity is O(n).