Add solution for customers who bought all products problem
Added a Python solution to report the customer IDs from the 'Customer' table that bought all the products in the 'Product' table. Return the result table in any order.

Method used: Group the 'customer' DataFrame by 'customer_id' and count the unique 'product_keys' for each customer. Filter customers who bought a number of unique products equal to the total number of products in the 'product' DataFrame.

Time complexity is O(n) and space complexity is O(n).