Add solution for invalid tweets problem (SQL)
Added an SQL script to create the Tweets table and insert sample data. The script also includes a query to find the IDs of the invalid tweets. A tweet is considered invalid if the content's character length is strictly greater than 15.

Method used: Use a SELECT statement to retrieve 'tweet_id' from the Tweets table where the LENGTH(content) is greater than 15.