Add solution for swap salary problem (SQL)
Added an SQL script to create the salary table and insert sample data. The script also includes a query to swap all 'f' and 'm' values in the 'gender' column of the salary table. A single UPDATE statement is used without any intermediate temporary tables.

Method used: Use an UPDATE statement with a CASE expression to change 'f' values to 'm' and vice versa in the 'gender' column.