Add solution for replace employee ID with the unique identifier problem (SQL)
Added an SQL script to create the Employees and EmployeeUNI tables and insert sample data. The script also includes a query to show the unique ID of each user. If a user does not have a unique ID in the EmployeeUNI table, it shows NULL.

Method used: Use a LEFT JOIN to combine the Employees and EmployeeUNI tables on the 'id' column. Select the 'unique_id' and 'name' columns from the joined tables.