Add solution for bank account summary II problem (SQL)
Added an SQL script to create the Users and Transactions tables and insert sample data. The script also includes a query to report the name and balance of users with a balance higher than 10000. The balance of an account is equal to the sum of the amounts of all transactions involving that account.

Method used: Use a SELECT statement to retrieve 'name' and the sum of 'amount' (aliased as 'balance') from the Users and Transactions tables. Join the tables on the 'account' column. Group the result by 'account' and filter using HAVING to include only those with a balance greater than 10000.